# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/plans/current": {
      "get": {
        "description": "The active plan, or null when the caller has none yet. `preparing` is set while a build that will replace (or create) the active plan is in progress, with F14 generation progress when it is running. `failure` is set when the newest such build failed, and is null while `preparing` is set.",
        "operationId": "PlansController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The caller's current plan.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CurrentPlanView"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH003: no valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Read the caller's current study plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/plans": {
      "get": {
        "description": "Up to the 100 newest plans, active and archived, newest first, each with its own completion statistics.",
        "operationId": "PlansController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The caller's plan history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlanHistoryView"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH003: no valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "List the caller's plans",
        "tags": [
          "plans"
        ]
      }
    },
    "/plans/{planId}": {
      "get": {
        "description": "Any of the caller's own plans, active or archived, by id.",
        "operationId": "PlansController_detail",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The plan.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/StudyPlanView"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VAL001: planId is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "AUTH003: no valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "PLAN001: no such plan for the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Read one of the caller's plans",
        "tags": [
          "plans"
        ]
      }
    },
    "/plans/retry": {
      "post": {
        "description": "Retries exactly the build `GET /plans/current` reports as `failure`. A `lesson`-origin failure is retried through the pipeline; a `recording_failed` or `analysis_blocked` failure is reset for the request job to pick up again.",
        "operationId": "PlansController_retryFailed",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The current plan view, now preparing.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CurrentPlanView"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH003: no valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "PLAN002: there is no failed plan build to retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionBearer": []
          },
          {
            "sessionCookie": []
          }
        ],
        "summary": "Retry the caller's failed plan build",
        "tags": [
          "plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "English Quest API",
    "description": "API for the English Quest study platform.\n\nEvery successful response nests its payload under `data`; every failure nests\nunder `error` with a stable `code`. Clients should switch on the code, never\non the message.\n\nAuthentication has two transports carrying the same opaque token. Browsers use a\n`eq_session` cookie, issued by `POST /auth/login` — HttpOnly and signed, so it\ncannot be read or forged from JavaScript. Clients with no cookie jar (the native\nmobile app) use `POST /auth/token` instead, which returns the token in the body\nand sets no cookie, and carry it as `Authorization: Bearer <token>` on every\nlater request. In Postman or Insomnia, enable the cookie jar and log in through\n`/auth/login`, or call `/auth/token` and set the bearer token on the collection.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "eq_session"
      },
      "sessionBearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque token from POST /auth/token",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTH001",
                  "AUTH002",
                  "AUTH003",
                  "AUTH004",
                  "CRED001",
                  "CRED002",
                  "CRED003",
                  "VAL001",
                  "PROMPT001",
                  "PROMPT002",
                  "PROMPT003",
                  "PROMPT004",
                  "CLASS001",
                  "CLASS002",
                  "CLASS003",
                  "CLASS004",
                  "SCEN001",
                  "SCEN002",
                  "SCEN003",
                  "REC001",
                  "REC002",
                  "PIPE001",
                  "PIPE002",
                  "PROF001",
                  "CONTENT001",
                  "CONTENT002",
                  "PLAN001",
                  "PLAN002",
                  "PLAN003",
                  "PLAN004",
                  "HEALTH001",
                  "ERR500"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "anyOf": [
                  {},
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "PlanHistoryView": {
        "type": "object",
        "properties": {
          "plans": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "active",
                    "archived"
                  ]
                },
                "origin": {
                  "type": "string",
                  "enum": [
                    "lesson",
                    "recording_failed",
                    "analysis_blocked"
                  ]
                },
                "lessonId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "lessonDate": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "activatedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "archivedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "summaryLine": {
                  "type": "string"
                },
                "activityCount": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "completedCount": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "skippedCount": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "completionPercent": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                },
                "ratings": {
                  "type": "object",
                  "properties": {
                    "tooEasy": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "justRight": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "tooHard": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "notUseful": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "tooEasy",
                    "justRight",
                    "tooHard",
                    "notUseful"
                  ],
                  "additionalProperties": false
                }
              },
              "required": [
                "id",
                "status",
                "origin",
                "lessonId",
                "lessonDate",
                "createdAt",
                "activatedAt",
                "archivedAt",
                "summaryLine",
                "activityCount",
                "completedCount",
                "skippedCount",
                "completionPercent",
                "ratings"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "plans"
        ],
        "additionalProperties": false
      }
    }
  }
}